new writer file Signed-off-by: Owen Diehl <ow.diehl@gmail.com>
